feature/EPMLCID-15564 fix upgrade sequence bug

+++ EPMLiveCore/EPMLiveCore/Jobs/EPMLiveUpgrade/Infrastructure/UpgradeUtilities.cs
@@ -21,11 +21,9 @@ namespace EPMLiveCore.Jobs.EPMLiveUpgrade.Infrastructure
 
         // Internal Methods (3) 
 
-        internal static Dictionary<double, Type> GetUpgradeSteps(string version)
+        internal static Dictionary<long, Type> GetUpgradeSteps(string version)
         {
-            var versionedSteps = new Dictionary<double, Type>();
-            //var genericSteps = new Dictionary<double, Type>();
-
+            var versionedSteps = new Dictionary<long, Type>();
             foreach (Type type in Assembly.GetExecutingAssembly().GetTypes())
             {
                 if (type.BaseType != typeof (UpgradeStep)) continue;
@@ -38,7 +36,7 @@ namespace EPMLiveCore.Jobs.EPMLiveUpgrade.Infrastructure
 
                 if (attribute.Version == EPMLiveVersion.GENERIC)
                 {
-                    versionedSteps.Add(0.0 - attribute.SequenceOrder, type);
+                    versionedSteps.Add(0 - attribute.SequenceOrder, type);
                 }
                 else
                 {
@@ -56,10 +54,7 @@ namespace EPMLiveCore.Jobs.EPMLiveUpgrade.Infrastructure
                     if (add) versionedSteps.Add(attribute.SequenceOrder, type);
                 }
             }
-
-            //List<Type> steps = versionedSteps.OrderBy(s => s.Key).Select(s => s.Value).ToList();
-            //steps.AddRange(genericSteps.OrderBy(s => s.Key).Select(source => source.Value));
-
+           
             return versionedSteps;
         }
